broken: 80% of work for lazy-loading refactor
this will eventually make it so *both* homework 6 and 9 modes lazy-load
their validation and blocking work, as part of the window that was
launched. That is -- be it parsing host ip, or requesting, waiting,
parsing username resolutions -- chat windows launch instantly and allow
user to interact once validations complete (giving back control of
launcher-dashboard immediately).